# sass/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MIS API 文档",
        "description": "伟明环保设备有限公司SIS系统 API 文档 api/documentation  php artisan l5-swagger:generate",
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/orgnizations/page": {
            "get": {
                "tags": [
                    "公司组织架构orgnizations"
                ],
                "summary": "分页获取电厂数据列表",
                "description": "使用说明：分页获取电厂数据列表",
                "operationId": "orgnizations-page",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "num",
                        "in": "query",
                        "description": "每页数据量",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "页数",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "中文名称搜索",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succeed"
                    }
                }
            }
        },
        "/api/orgnizations/switch": {
            "post": {
                "tags": [
                    "公司组织架构orgnizations"
                ],
                "summary": "切换组织",
                "description": "使用说明：切换组织",
                "operationId": "orgnizations-switch",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "组织ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succeed"
                    }
                }
            }
        },
        "/api/orgnizations/factories": {
            "get": {
                "tags": [
                    "公司组织架构orgnizations"
                ],
                "summary": "获取所有电厂列表",
                "description": "使用说明：获取所有电厂列表",
                "operationId": "orgnizations-factories",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succeed"
                    }
                }
            }
        },
        "/api/orgnizations/tree": {
            "get": {
                "tags": [
                    "公司组织架构orgnizations"
                ],
                "summary": "组织树",
                "description": "使用说明：获取组织树",
                "operationId": "orgnizationsAll",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "level",
                        "in": "query",
                        "description": "最大层级",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/orgnizations/store": {
            "post": {
                "tags": [
                    "公司组织架构orgnizations"
                ],
                "summary": "保存组织",
                "description": "使用说明：保存组织",
                "operationId": "storeOrgnization",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "组织id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "组织名称",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "组织编码(电厂英文全拼+数字)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "组织描述",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "parent_id",
                        "in": "query",
                        "description": "父组织ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "排序号",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/orgnizations/{id}/role": {
            "get": {
                "tags": [
                    "公司组织架构orgnizations"
                ],
                "summary": "组织角色",
                "description": "使用说明：获取组织角色",
                "operationId": "getOrgRole",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "用户组织ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "公司组织架构orgnizations"
                ],
                "summary": "保存组织角色",
                "description": "使用说明：保存组织角色",
                "operationId": "storeOrgRole",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "用户组织ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "roles",
                        "in": "query",
                        "description": "角色ID列表",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/orgnizations/delete": {
            "post": {
                "tags": [
                    "公司组织架构orgnizations"
                ],
                "summary": "删除组织",
                "description": "使用说明：删除组织",
                "operationId": "deleteOrgnization",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "用户组织ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
...
}